refactor: move ggml format to module
