+++ crates/ggml/src/format/ggml/loader.rs
@@ -6,67 +6,16 @@
 
 use std::{
     error::Error,
-    fmt,
     io::{BufRead, Seek, SeekFrom},
 };
 
 use crate::{
+    format::{data_size, header_size, LoadError},
     util::{has_data_left, read_bytes_with_len, read_f32, read_i32, read_u32},
-    ContainerType, ElementType,
+    ElementType,
 };
 
-/// Helper struct that wraps the magic number of a file format,
-/// so that it can be printed in a human-readable format.
-pub struct FormatMagic(pub u32);
-impl fmt::Display for FormatMagic {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{:x} ({})",
-            self.0,
-            String::from_utf8_lossy(&self.0.to_le_bytes())
-        )
-    }
-}
-impl fmt::Debug for FormatMagic {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> std::fmt::Result {
-        fmt::Display::fmt(self, f)
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-/// Errors that can occur while loading a model.
-pub enum LoadError<E: Error> {
-    #[error("invalid file magic number: {0}")]
-    /// The file magic number is invalid.
-    InvalidMagic(FormatMagic),
-    #[error("invalid ggml format: format={0:?}")]
-    /// An unsupported format version was found.
-    InvalidFormatVersion(ContainerType),
-    #[error("non-specific I/O error")]
-    /// A non-specific IO error.
-    Io(#[from] std::io::Error),
-    #[error("could not convert bytes to a UTF-8 string")]
-    /// One of the strings encountered was not valid UTF-8.
-    InvalidUtf8(#[from] std::string::FromUtf8Error),
-    #[error("invalid integer conversion")]
-    /// One of the integers encountered could not be converted to a more appropriate type.
-    InvalidIntegerConversion(#[from] std::num::TryFromIntError),
-    #[error("implementation error")]
-    /// An error `E` was returned by the implementation of the loader.
-    ImplementationError(#[source] E),
-    #[error("unsupported tensor type {ftype} for tensor {tensor_name}")]
-    /// One of the tensors encountered had an unsupported data type.
-    UnsupportedElementType {
-        /// The name of the tensor.
-        tensor_name: String,
-        /// The format type that was encountered.
-        ftype: u32,
-    },
-    #[error("invariant broken: {0}")]
-    /// An invariant was broken.
-    InvariantBroken(String),
-}
+use super::ContainerType;
 
 #[derive(Debug, Clone)]
 /// Information about a [tensor](https://en.wikipedia.org/wiki/Tensor_(machine_learning)) that is being read.
@@ -116,21 +65,6 @@ impl TensorLoadInfo {
         reader.read_exact(&mut data)?;
         Ok(data)
     }
-}
-
-/// Returns the size occupied by a tensor's data in bytes given the element type and number of elements.
-pub(crate) fn data_size(element_type: ElementType, n_elements: usize) -> usize {
-    (crate::type_size(element_type) * n_elements) / crate::blck_size(element_type)
-}
-
-/// Returns the size of the ggml tensor header in bytes.
-pub(crate) fn header_size() -> usize {
-    crate::Tensor::C_TYPE_SIZE + crate::OBJECT_SIZE
-}
-
-/// Returns the size of a tensor in bytes given the element type and number of elements. This includes the tensor's header.
-pub fn tensor_size(element_type: ElementType, n_elements: usize) -> usize {
-    header_size() + data_size(element_type, n_elements)
 }
 
 #[derive(Debug, Clone)]
